+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # telepot changelog
 
+## 9.1 (2016-08-26)
+
+- Changed the name `pave_callback_query_origin_map()` to `intercept_callback_query_origin()`
+- Added `include_callback_query_chat_id()`
+
+## 9.0 (2016-08-25)
+
+- I am finally satisfied with callback query handling. Many styles of dealing with
+callback query are now possible.
+- Added a few `per_callback_query_*()` seeder factories
+- Added a few pair producers, e.g. `pave_event_space()`, `pave_callback_query_origin_map()`
+- Added `Bot.Scheduler` to schedule internal events
+- Invented a standard event format for delegates to create their own events easily
+- Improved Mixin framework. Added `StandardEventMixin`, `IdleTerminateMixin`, and
+`InterceptCallbackQueryMixin`.
+- Added `CallbackQueryOriginHandler`
+- Revamped `Listener` and message capture specifications
+- Default `retries=3` for `urllib3`
+- Relaxed `urllib3>=1.9.1` in `setup.py`
+
+## 8.3 (2016-07-17)
+
+- Fixed `urllib3==1.9.1` in `setup.py`
+
 ## 8.2 (2016-07-04)
 
 - Handling of callback query still unsatisfactory, a transitional release